implement highlight code and rasterize code
I don't think `rasterize` is the correct name here, but the idea is to
split the string of code into a 2D array with each character per "cell".

We also make sure that each character is colored on its own, so that
injecting any other information in the array will have the default text
color, background color and styles applied.

This will also allow us to have more control over the data, where we can
use this 2D array and give each cell a type and some associated
information that we can use later in the code. At the end, we can then
print it all back to a single string when flushing to the terminal.