rework notes
Reworked the notes completely and now the `notes` are either not passed
or they are a single string.

The string can contain markdown that we render as best as we can. The
nice part is that you get a lot of free-form space where you can note
down anything that's important to the issue. From links to
documentation, to actual suggestions with codeblocks.